Compact InputSection from 64 to 56 bytes (shndx instead of a header pointer)
The subsection held an 8-byte pointer to its parent section's header.
mold-rust's InputSection holds a u32 shndx and resolves the header
through its file; do the same. Context::hdr_of(&isec) looks the header
up in the owning object's section list, or in a small synthetic-header
table for the sections the linker makes itself (the __common section
for tentative definitions). The subsection's size becomes a u32 as
well - one subsection is far under 4 GiB - which with the 4 bytes from
the header index crosses the alignment boundary: 62 -> 54 bytes of
fields, 64 -> 56 after alignment, asserted at compile time. That is 8
more bytes off the ~4.4M-element arena of a debug clang link, and
InputSection is now smaller than mold-rust's (it has no section-name
or flags word; we keep the compact-unwind range instead).

The output-section assignment loop binds the header once per
subsection before it takes the chunk mutably (the same pointer cache as
before). 322 integration tests pass; the debug clang, nu and mold links
run.

Interleaved A/B against the parent commit, debug links, best 4 of 8
runs (ms): clang 778.1 -> 785.1 (+0.9%), nu 370.0 -> 371.6 (+0.4%),
mold 148.6 -> 153.5 (+3.3%); an earlier run of the same pair read
-0.7%, +0.9%, +1.1% - within noise either way. The gain is the 8 bytes
per subsection, not time.

+++ src/mapfile.rs
@@ -114,7 +114,7 @@ pub fn print_map<E: Arch>(ctx: &Context<E>) {
         let (_, _, _, isec, value) = syms[idx];
         let end = match order.get(i + 1) {
             Some(&next) if syms[next].3 == isec => syms[next].4,
-            _ => ctx.isecs[isec].size,
+            _ => ctx.isecs[isec].size as u64,
         };
         sizes[idx] = end.saturating_sub(value);
     }
@@ -163,7 +163,7 @@ pub fn print_map<E: Arch>(ctx: &Context<E>) {
         for (i, &(file, value, isec, name)) in dead.iter().enumerate() {
             let end = match dead.get(i + 1) {
                 Some(&(_, next_value, next_isec, _)) if next_isec == isec => next_value,
-                _ => ctx.isecs[isec].size,
+                _ => ctx.isecs[isec].size as u64,
             };
             let _ = writeln!(
                 out,
